fixed modal who asked if sure of deleting

+++ src/components/CardRecipe.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 export const CardRecipe = ({
   id,
@@ -12,13 +13,27 @@ export const CardRecipe = ({
 }) => {
   const maxLength = 180;
 
-  const myData = { id, title, elaboration, ingredients, imgUrl, deleteRecipe };
+  const myData = { id, title, elaboration, ingredients, imgUrl };
 
   const noimg =
     "https://upload.wikimedia.org/wikipedia/commons/thumb/a/ac/No_image_available.svg/480px-No_image_available.svg.png";
 
   const handleDelete = () => {
-    deleteRecipe(id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#5c9bdb",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Deleted!", "Your file has been deleted.", "success");
+
+        deleteRecipe(id);
+      }
+    });
   };
 
   return (
